+++ R/adjustedcostbase.ca_1.R
@@ -6,11 +6,27 @@
 #' @docType data
 #' @format A data frame with 4 rows and 4 variables:
 #' \describe{
+#'   \item{date}{the date}
 #'   \item{transaction}{buy or sell}
-#'   \item{price}{in dollars}
 #'   \item{quantity}{how much of the stock/coin}
+#'   \item{price}{the spot rate, in dollars}
 #'   \item{fees}{any transaction fees}
 #'   ...
 #' }
 #' @source \url{https://www.adjustedcostbase.ca/blog/how-to-calculate-adjusted-cost-base-acb-and-capital-gains/}
-"adjustedcostbase.ca"
+"adjustedcostbase.ca_1"
+
+#' @rdname adjustedcostbase.ca_1
+"adjustedcostbase.ca_2"
+
+#' @rdname adjustedcostbase.ca_1
+"adjustedcostbase.ca_3"
+
+#' @rdname adjustedcostbase.ca_1
+"adjustedcostbase.ca_4"
+
+#' @rdname adjustedcostbase.ca_1
+"adjustedcostbase.ca_5"
+
+#' @rdname adjustedcostbase.ca_1
+"adjustedcostbase.ca_6"