+++ R/format_CDC_exchange_rewards.R
@@ -1,14 +1,23 @@
 #' @title Format CDC exchange file (FOR REWARDS ONLY)
 #'
-#' @description Format a .csv transaction history file from the Crypto.com exchange for later ACB processing. Only processes rewards, not trades (see `format_CDC_exchange_trades` for this).
+#' @description Format a .csv transaction history file from the
+#' Crypto.com exchange for later ACB processing. Only processes
+#' rewards and withdrawal fees, not trades (see
+#' `format_CDC_exchange_trades` for this).
 #'
-#' To download the rewards data from the Crypto.com exchange as a CSV file, copy and paste the code below and save it as a bookmark in your browser.
+#' To download the rewards/withdrawal fees data from the Crypto.com
+#' exchange as a CSV file, copy and paste the code below and save it
+#' as a bookmark in your browser.
 #'
 #' `javascript:(function(){function callback(){window.cdc()}var s=document.createElement("script");s.src="https://cdn.jsdelivr.net/gh/ConorIA/cdc-csv@master/cdc.js";if(s.addEventListener){s.addEventListener("load",callback,false)}else if(s.readyState){s.onreadystatechange=callback}document.body.appendChild(s);})()`
 #'
-#' Then log into the crypto.com exchange and click the bookmark you saved. It will automatically download a CSV that contains Supercharger rewards, withdrawal fees, CRO staking interest (if you have an exchange stake), among others.
+#' Then log into the crypto.com exchange and click the bookmark you
+#' saved. It will automatically download a CSV that contains Supercharger
+#' rewards, withdrawal fees, CRO staking interest (if you have an
+#' exchange stake), among others.
 #'
-#' Note that this code does not include the initial referral reward in CRO for signup or on the Crypto.com exchange. It must be added manually.
+#' Note that this code does not include the initial referral reward in CRO
+#' for signup or on the Crypto.com exchange. It must be added manually.
 #'
 #' WARNING: DOES NOT DOWNLOAD TRADES, ONLY REWARDS!
 #'
@@ -80,8 +89,20 @@ format_CDC_exchange_rewards <- function(data) {
       "revenue.type", "description", "comment"
     )
 
+  # Create a "withdrawals" object
+  WITHDRAWALS <- data %>%
+    filter(.data$description == "Withdrawal") %>%
+    mutate(
+      quantity = .data$Fee.Amount,
+      transaction = "sell"
+    ) %>%
+    select(
+      "date", "quantity", "currency", "transaction",
+      "description"
+    )
+
   # Merge the "buy" and "sell" objects
-  data <- EARN %>%
+  data <- bind_rows(EARN, WITHDRAWALS) %>%
     mutate(
       fees = 0,
       exchange = "CDC.exchange"