+++ R/format_binance_withdrawals.R
@@ -1,8 +1,16 @@
 #' @title Format Binance withdrawal file
 #'
-#' @description Format a .xlsx withdrawal history file from Binance for later ACB processing.
+#' @description Format a .xlsx withdrawal history file from Binance for later
+#' ACB processing.
+#' @details To get this file, connect to your Binance account on
+#' desktop, click "Wallet" (top right), "Transaction History", then in the
+#' "Type" column, choose "Withdraw". Next, click on "Export Withdrawal History"
+#' on the right and choose your time frame (you will probably need to choose
+#' "Customized"). You are only allowed to choose up to 3 months, so you might
+#' have to download more than one file and merge them before using this function.
 #'
-#' Warning: This does NOT process TRADES See the format_binance_trades() function for this purpose.
+#' Warning: This does NOT process TRADES See the `format_binance_trades()`
+#' function for this purpose.
 #' @param data The dataframe
 #' @keywords money crypto
 #' @export